Added use of cryptographic library - voting key generation, counting of voting results

+++ src/main/java/edu/karazin/secure_voting_node/dto/VoteResultDto.java
@@ -8,5 +8,5 @@ import java.util.Map;
 @Data
 @AllArgsConstructor
 public class VoteResultDto {
-    private Map<String, Integer> results;
+    private Map<String, Long> results;
 }
